requirements.txt

# Web Scraping & Data Collection
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
selenium==4.16.0
PyPDF2==3.0.1
pdfplumber==0.10.4

# Data Processing
pandas==2.2.0
numpy==1.26.3
tqdm==4.66.1

# Text Processing
nltk==3.8.1
spacy==3.7.2

# Document Processing
python-docx==1.1.0
openpyxl==3.1.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyYAML==6.0.1

# API & Web Framework (optional, for future API endpoint)
fastapi==0.109.0
uvicorn==0.27.0

# Logging & Monitoring
loguru==0.7.2
